requirements.txt
aiohttp~=3.9.3
beautifulsoup4~=4.12.3
discord.py[voice]~=2.3.2
gitpython~=3.1.42
lxml~=5.1.0
pandas~=2.2.1
Pillow~=10.2.0
requests~=2.31.0
toml~=0.10.2
Unidecode~=1.3.8
sqlalchemy~=2.0.28
wavelink~=3.2.0
psycopg2-binary~=2.9.9
cryptography~=42.0.5
